# Conflicts:
#	XF.Material/UI/MaterialTextField.xaml

+++ XF.Material/UI/MaterialElevation.cs
@@ -7,19 +7,19 @@ namespace XF.Material.Forms.UI
     [TypeConverter(typeof(MaterialElevationTypeConverter))]
     public struct MaterialElevation
     {
-        public int RestingElevation { get; }
+        public double RestingElevation { get; }
 
-        public int PressedElevation { get; }
+        public double PressedElevation { get; }
 
-        public MaterialElevation(int uniformElevation) : this(uniformElevation, uniformElevation) { }
+        public MaterialElevation(double uniformElevation) : this(uniformElevation, uniformElevation) { }
 
-        public MaterialElevation(int restingElevation, int pressedElevation)
+        public MaterialElevation(double restingElevation, double pressedElevation)
         {
             RestingElevation = restingElevation;
             PressedElevation = pressedElevation;
         }
 
-        public static implicit operator MaterialElevation(int uniformElevation) => new MaterialElevation(uniformElevation);
+        public static implicit operator MaterialElevation(double uniformElevation) => new MaterialElevation(uniformElevation);
     }
 
     [Xamarin.Forms.Xaml.TypeConversion(typeof(MaterialElevation))]
@@ -41,14 +41,14 @@ namespace XF.Material.Forms.UI
                 switch (elevations.Length)
                 {
                     case 1:
-                        if (int.TryParse(elevations[0], NumberStyles.Number, CultureInfo.InvariantCulture, out var uE))
+                        if (double.TryParse(elevations[0], NumberStyles.Number, CultureInfo.InvariantCulture, out var uE))
                         {
                             return new MaterialElevation(uE);
                         }
                         break;
                     case 2:
-                        if (int.TryParse(elevations[0], NumberStyles.Number, CultureInfo.InvariantCulture, out var rE)
-                            && int.TryParse(elevations[1], NumberStyles.Number, CultureInfo.InvariantCulture, out var pE))
+                        if (double.TryParse(elevations[0], NumberStyles.Number, CultureInfo.InvariantCulture, out var rE)
+                            && double.TryParse(elevations[1], NumberStyles.Number, CultureInfo.InvariantCulture, out var pE))
                         {
                             return new MaterialElevation(rE, pE);
                         }
@@ -57,7 +57,7 @@ namespace XF.Material.Forms.UI
                         throw new InvalidOperationException($"Cannot convert {value} to {typeof(MaterialElevation)}");
                 }
             }
-            else if (int.TryParse(value, NumberStyles.Number, CultureInfo.InvariantCulture, out var uE))
+            else if (double.TryParse(value, NumberStyles.Number, CultureInfo.InvariantCulture, out var uE))
             {
                 return new MaterialElevation(uE);
             }
